% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_values.R
\name{extract_values}
\alias{extract_values}
\title{Extract values of input raster data in the buffered data}
\usage{
extract_values(buffered_data, input_raster, target_variable)
}
\arguments{
\item{buffered_data}{sf object with buffered data}

\item{input_raster}{raster object for extraction}

\item{target_variable}{character string for the target variable (min, max, mean, median, etc.). See exact_extract}
}
\value{
sf object with summary statistics
}
\description{
Extract values of input raster data in the buffered data
}
