% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_alignment.R
\name{linear_alignment}
\alias{linear_alignment}
\title{Linear Alignment of Points}
\usage{
linear_alignment(sf_object, crs_code)
}
\arguments{
\item{sf_object}{An sf object containing spatial features.}

\item{crs_code}{An integer representing the coordinate reference system (CRS) code to transform the sf object.}
}
\value{
An sf object with adjusted coordinates that best fit a line.
}
\description{
This function aligns satellite LiDAR footprints to a best fit line using genetic algorithm optimization.
The sf object should be transformed to the specified (projected) CRS before using this function.
The fitting is done by minimizing the perpendicular distances between the points and the line using a genetic algorithm (GA) optimization.
}
\examples{
\dontrun{
library(sf)
library(GA)

# Example usage
sf_object <- st_read("path_to_your_shapefile.shp")
crs_code <- 3857
adjusted_sf <- linear_alignment(sf_object, crs_code)
}

}
