% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_correction.R
\name{location_correction}
\alias{location_correction}
\title{Correct the location of the data}
\usage{
location_correction(gedi_l2a, add_x, add_y, buf, crs_code)
}
\arguments{
\item{gedi_l2a}{sf object of gedi l2a}

\item{add_x}{numeric value to add to x coordinates}

\item{add_y}{numeric value to add to y coordinates}

\item{buf}{numeric value for buffer size}

\item{crs_code}{numeric value for CRS code}
}
\value{
sf object with corrected location
}
\description{
Correct the location of the data
}
