% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_dtw.R
\name{perform_dtw}
\alias{perform_dtw}
\title{Perform Dynamic Time Warping (DTW) on two sequences}
\usage{
perform_dtw(seq1, seq2)
}
\arguments{
\item{seq1}{A numeric vector representing the first sequence.}

\item{seq2}{A numeric vector representing the second sequence.}
}
\value{
A numeric value representing the normalized DTW alignment distance between the two sequences.
}
\description{
This function takes two sequences, removes any rows with NA values, and performs
Dynamic Time Warping (DTW) to calculate the alignment distance between the cleaned sequences.
The alignment distance is then normalized by the length of the first sequence.
}
\examples{
seq1 <- c(1, 2, 3, 4, 5)
seq2 <- c(2, 3, 4, 5, 6)
perform_dtw(seq1, seq2)
}
