% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_adjustment.R
\name{position_adjustment}
\alias{position_adjustment}
\title{Positional adjustment of satellite LiDAR footprints}
\usage{
position_adjustment(lidar_footprints, add_x, add_y, crs_code)
}
\arguments{
\item{lidar_footprints}{Input satellite LiDAR footprints as an sf object}

\item{add_x}{X-axis offset}

\item{add_y}{Y-axis offset}

\item{crs_code}{Coordinate reference system code}
}
\value{
Adjusted satellite LiDAR footprints as an sf object with updated geometry
}
\description{
This function adjusts the point locations of satellite LiDAR footprints with specified offsets.
}
