% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positional_correction.R
\name{positional_correction}
\alias{positional_correction}
\title{Positional adjustment of satellite LiDAR footprints}
\usage{
positional_correction(lidar_footprints, add_x, add_y, buf, crs_code)
}
\arguments{
\item{lidar_footprints}{Input satellite LiDAR footprints as an sf object}

\item{add_x}{X-axis offset}

\item{add_y}{Y-axis offset}

\item{buf}{Buffer size}

\item{crs_code}{Coordinate reference system code}
}
\value{
Adjusted satellite LiDAR footprints
}
\description{
This function adjusts the point locations of satellite LiDAR footprints with specified offsets.
}
