% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_loss.R
\name{get_loss}
\alias{get_loss}
\title{Get loss value to be minimized}
\usage{
get_loss(
  gedi_l2a,
  add_x,
  add_y,
  buf,
  input_raster,
  minimizing_method,
  target_variable,
  crs_code,
  gedi_value
)
}
\arguments{
\item{gedi_l2a}{sf object with GEDI L2A data}

\item{add_x}{numeric value to add to x coordinates}

\item{add_y}{numeric value to add to y coordinates}

\item{buf}{numeric value for buffer size}

\item{input_raster}{raster object for extraction}

\item{minimizing_method}{character string for the minimizing method}

\item{target_variable}{character string for the target variable}

\item{crs_code}{numeric value for CRS code}

\item{gedi_value}{numeric value for GEDI data}
}
\value{
data frame with summary statistics
}
\description{
Get loss value to be minimized
}
