% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_alignment.R
\name{linear_alignment}
\alias{linear_alignment}
\title{Linear Alignment of Points}
\usage{
linear_alignment(sf_object, crs_code)
}
\arguments{
\item{sf_object}{An sf object containing spatial features. It must include a column named 'shot_number'.}

\item{crs_code}{An integer representing the coordinate reference system (CRS) code to which the sf object will be transformed.}
}
\value{
An sf object with adjusted coordinates that best fit a line, including original coordinates stored in 'orig_x' and 'orig_y'.
}
\description{
This function aligns satellite LiDAR footprints to a best-fit line using an analytical total least squares approach.
The method computes the principal component of the coordinates for each shot group to minimize the perpendicular distances
between the points and the line, and then projects the points onto this line.
}
\examples{
\dontrun{
library(sf)

# Example usage
sf_object <- st_read('path_to_your_shapefile.shp')
crs_code <- 3857
adjusted_sf <- linear_alignment(sf_object, crs_code)
}

}
