% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimize_loss.R
\name{minimize_loss}
\alias{minimize_loss}
\title{Minimize loss}
\usage{
minimize_loss(
  lidar_footprints,
  input_rast,
  minimizing_method,
  target_variable,
  buf,
  crs_code,
  lidar_value,
  lower_bounds,
  upper_bounds,
  pop_size,
  max_iter,
  parallel
)
}
\arguments{
\item{lidar_footprints}{sf object containing satellite LiDAR footprints}

\item{input_rast}{referenced raster object for extraction}

\item{minimizing_method}{character string for the minimizing method ("dtw" is only supported)}

\item{target_variable}{character string for the target variable (min, max, mean, median, etc.). See exact_extract}

\item{buf}{numeric value for buf size. Default is 12.5}

\item{crs_code}{numeric value for CRS code}

\item{lidar_value}{name of the column containing LiDAR measurement values}

\item{lower_bounds}{numeric vector for lower bounds in GA. Default is c(-30, -30)}

\item{upper_bounds}{numeric vector for upper bounds in GA. Default is c(30, 30)}

\item{pop_size}{numeric value for population size in GA. Default is 50}

\item{max_iter}{numeric value for maximum iterations in GA. Default is 100}

\item{parallel}{logical value for parallel processing. Default is FALSE}
}
\value{
list with optimization results
}
\description{
Minimize loss
}
