% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positional_correction.R
\name{positional_correction}
\alias{positional_correction}
\title{Positional correction of satellite LiDAR footprints}
\usage{
positional_correction(
  lidar_footprints,
  input_rast,
  minimizing_method = "euclidean",
  target_variable,
  buf = 12.5,
  crs_code,
  lidar_value,
  optimization_method = "ga",
  lower_bounds = c(-25, -25),
  upper_bounds = c(25, 25),
  pop_size = 50,
  max_iter = 100,
  parallel = FALSE,
  pso_params = NULL,
  geographic_coordinates = FALSE
)
}
\arguments{
\item{lidar_footprints}{sf object containing satellite LiDAR footprints.}

\item{input_rast}{Raster object for reference data.}

\item{minimizing_method}{Character string for the minimizing method. Default is "euclidean".
Options include "dtw", "euclidean", "manhattan", "correlation", "area", "hausdorff".}

\item{target_variable}{Character string for the target variable.}

\item{buf}{Numeric value for buffer size (default: 12.5).}

\item{crs_code}{Numeric value for CRS code.}

\item{lidar_value}{Name of the column containing LiDAR measurement values.}

\item{optimization_method}{Character string for optimization method. Options are "ga" (genetic algorithm) or "pso" (particle swarm optimization). Default is "ga".}

\item{lower_bounds}{Numeric vector for lower bounds in the optimization (default: c(-30, -30)).}

\item{upper_bounds}{Numeric vector for upper bounds in the optimization (default: c(30, 30)).}

\item{pop_size}{Numeric value for population size (GA) or swarm size (PSO) (default: 50).}

\item{max_iter}{Numeric value for maximum iterations (default: 100).}

\item{parallel}{Logical value for parallel processing (default: FALSE).}

\item{pso_params}{List of additional parameters for PSO (default: NULL).
List can include: w (inertia), c1 (cognitive learning factor), c2 (social learning factor).}

\item{geographic_coordinates}{Logical value indicating if input coordinates are in geographic coordinates (default: FALSE).
If TRUE, the bounds will be converted from meters to degrees.}
}
\value{
List containing optimization results and corrected positions.
}
\description{
This function performs positional correction on satellite LiDAR footprints using optimization algorithms.
}
