% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis.R
\name{run_analysis}
\alias{run_analysis}
\title{Run the analysis for all orbits}
\usage{
run_analysis(
  lidar_footprints,
  input_raster,
  minimizing_method = "dtw",
  target_variable,
  buf = 12.5,
  crs_code,
  lidar_value,
  lower_bounds = c(-30, -30),
  upper_bounds = c(30, 30),
  pop_size = 50,
  max_iter = 100,
  parallel = parallel
)
}
\arguments{
\item{lidar_footprints}{sf object containing satellite LiDAR footprints}

\item{input_raster}{raster object for extraction}

\item{minimizing_method}{character string for the minimizing method}

\item{target_variable}{character string for the target variable}

\item{buf}{numeric value for buffer size}

\item{crs_code}{numeric value for CRS code}

\item{lidar_value}{name of the column containing LiDAR measurement values}

\item{lower_bounds}{numeric vector for lower bounds in GA}

\item{upper_bounds}{numeric vector for upper bounds in GA}

\item{pop_size}{numeric value for population size in GA}

\item{max_iter}{numeric value for maximum iterations in GA}

\item{parallel}{logical value for parallel processing}
}
\value{
list with results for all orbits
}
\description{
Run the analysis for all orbits
}
